6 is done

+++ app/src/main/java/uk/edu/le/co2124/part2/database/entity/CourseStudentCrossRef.java
@@ -5,5 +5,5 @@ import androidx.room.Entity;
 @Entity(primaryKeys = {"courseId", "studentId"})
 public class CourseStudentCrossRef {
     public int courseId;
-    public int studentId;
+    public long studentId;
 }
